+++ app/Data/CrawlDataMetrics.php
@@ -2,7 +2,6 @@
 
 namespace App\Data;
 
-use App\Data\Factories\CrawlDataTimings;
 use Spatie\LaravelData\Data;
 use Spatie\LaravelData\DataCollection;
